graphql: fix resource states
The current code passes `initialValue` unconditionally to the
`createResource()` function, forcing an initial value even if
the corresponding argument is undefined. This is bad because
the sequence of states the resource goes through is:

* refreshing
* ready

while with an undefined `initialValue` the sequence is
expected to be:

* pending
* ready

This patch fixes this problem and restores the correct resource
states by passing a third argument to `createResource()` only
when `initialState !== undefined`.