Update index.ts - "smart" option to normalize returned values
Added new config option "smart"
If true all values will be returned as reactive functions, meaning it will resolve values of type MaybeAccessor  to Accessor.
This helps a lot for library devs to provide a flexible interface to users and don't worry about having to make breaking changes later on and the handling of the destructured props is always as Accessor.
About the naming feel free to rename it to anything you like.
I was thinking about:
normalize
normalizeValues
accessibleValues
...
Maybe it would also make sense to make this the default behavior, but that would be a breaking change.