Use `number` as the type for the tweened value
In future, it would be neat to include an `interpolate` prop to this primitive, which would allow it to be used to interpolate between other values like strings and colours. Until then, it is simpler to restrict the signal type that can be tweened to just numbers. This clear up a couple of TypeScript errors without having to recast values.

In addition, added a TweenProps type.